Put subnet addresses at start of connection lines
As each line giving a connection contains exactly one subnet and one or
more router/link pairs, it is easier to parse if the subnet address
comes first, rather than last. Since this has similar readability and
since the intermediate format is not used by an program not used
internally, this change is preferable.